requirements.txt
fastapi==0.79.0
uvicorn>=0.15.0,<0.16.0
urllib3==1.26.5
pillow==9.2.0
aiofiles
matplotlib==3.5.2	
git+https://github.com/openai/CLIP.git
torch==1.12.1	
jinja2>=3.0
starlette==0.19.1
python-multipart